Refactors imaging paddm wrapper
Rather than try to tie in the functionality as a derived class,
ImagingProximalADMM is now a wrapper. The cost is that all the data is
reproduced in the wrapper. The benefit is that are: (i)
ImagingProximalADMM is much more obviously just a version of PADMM with
some of the input specialized or reworked, (ii) removes the spagetthi
code that use to tie one input format to the other.