Test runs but solution does not converge.
Issues with conflicting data types should be resolved.
There is an error in the model call if I instantiate the model in the
constructor. I left some commented out code for debugging this,
currently the model is read on every model call which is extremely
slow and definitely not how we want to do it.
Because the data type of the image is t_Vector (instead of Image) the
functions in cppflowutils did not work directly. I overloaded the
preprocessing function but did not yet do anything with the
postprocessing function.